This aparently was broken, and is working now :)

+++ src/commands/Moderation/Reload.js
@@ -17,48 +17,50 @@ module.exports = {
             message.reply("Você não passou comando algum para recarregar.");
         }
 
-        const commandName = args[0].toLowerCase();
-        const command =
+        if (args.length) {    
+            const commandName = args[0].toLowerCase();
+            const command =
             message.client.commands.get(commandName) ||
             message.client.commands.find(
                 (cmd) => cmd.aliases && cmd.aliases.includes(commandName)
-            );
-
-        if (!command) {
-            message.reply(`Não tem comando com o nome \`${commandName}\`.`);
-        }
-        if (command) {
-            const commandFolders = fs.readdirSync("./commands");
-            
-            let folderName;
-            
-            commandFolders.map((folder) => {
-                fs.readdirSync(`${__dirname}/../${folder}/`).find((element) => {
-                    if (element === `${command.name}.js`) {
-                        delete require.cache[
-                            require.resolve(`../${folder}/${command.name}.js`)
-                        ];
-                        
-                        folderName = folder;
-                    }
-                });
-            });
-            
-            try {
-                const newCommand = require(`../${folderName}/${command.name}.js`);
-                message.client.commands.set(newCommand.name, newCommand);
-                console.log("\x1b[35m%s\x1b[0m", `Command ${newCommand.name} was reloaded!`);
-            } catch (error) {
-                console.error(error);
-                message.channel.send(
-                    `There was an error while reloading a command \`${command.name}\`:\n\`${error.message}\``
-                    );
-                }
-            if (message.channel.type === 'dm') {
-                return;
+                );        
+                
+            if (!command) {
+                message.reply(`Não tem comando com o nome \`${commandName}\`.`);
             }
-            else if (message.channel.type === 'text') {
-                message.delete();
+            if (command) {
+                const commandFolders = fs.readdirSync("./commands");
+                
+                let folderName;
+                
+                commandFolders.map((folder) => {
+                    fs.readdirSync(`${__dirname}/../${folder}/`).find((element) => {
+                        if (element === `${command.name}.js`) {
+                            delete require.cache[
+                                require.resolve(`../${folder}/${command.name}.js`)
+                            ];
+                            
+                            folderName = folder;
+                        }
+                    });
+                });
+                
+                try {
+                    const newCommand = require(`../${folderName}/${command.name}.js`);
+                    message.client.commands.set(newCommand.name, newCommand);
+                    console.log("\x1b[35m%s\x1b[0m", `Command ${newCommand.name} was reloaded!`);
+                } catch (error) {
+                    console.error(error);
+                    message.channel.send(
+                        `There was an error while reloading a command \`${command.name}\`:\n\`${error.message}\``
+                        );
+                }
+                if (message.channel.type === 'dm') {
+                    return;
+                }
+                else if (message.channel.type === 'text') {
+                    message.delete();
+                }
             }
         }
     },
